+++ WireMailPHPMailer.module.php
@@ -1,11 +1,11 @@
-<?php namespace ProcessWire;
+<?php
 
-// Import PHPMailer classes into the global namespace
-// These must be at the top of your script, not inside a function
+namespace ProcessWire;
+
+
 use PHPMailer\PHPMailer\PHPMailer;
 use PHPMailer\PHPMailer\Exception;
 
-//Load composer's autoloader
 require __DIR__ . DIRECTORY_SEPARATOR . "vendor/autoload.php";
 
 /**
@@ -15,19 +15,40 @@ require __DIR__ . DIRECTORY_SEPARATOR . "vendor/autoload.php";
  * @website			: https://www.altivebir.com
  * @projectWebsite	: https://github.com/trk/WireMailPHPMailer
  */
-class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule {
+class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule
+{
 
     /**
      * PHPMailer Version
      */
-    const PHPMailer_VERSION = "6.1.6";
+    const PHPMailer_VERSION = "6.1.8";
 
     /**
-     * PHPMailer instance
-     *
      * @var PHPMailer
      */
-    protected $PHPMailer;
+    protected $instance;
+
+    /**
+     * @var array
+     */
+    protected $options = [
+        'AltBody'   => '',
+        'Subject'   => '',
+        'setFrom'   => [
+            'address' => '',
+            'name' => '',
+            'auto' => ''
+        ],
+        'addAddress' => [],
+        'addCC' => [],
+        'addBCC' => [],
+        'addReplyTo' => [],
+        'msgHTML' => [
+            'message' => '',
+            'basedir' => '',
+            'advanced' => ''
+        ]
+    ];
 
     /**
      * Should we throw external exceptions?
@@ -44,10 +65,11 @@ class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule {
      * @see Module
      * @return array
      */
-    public static function getModuleInfo() {
+    public static function getModuleInfo()
+    {
         return array(
             'title' => 'WireMailPHPMailer',
-            'version' => 122,
+            'version' => 123,
             'summary' => __('This module extends WireMail base class, integrating the PHPMailer mailing library into ProcessWire.'),
             'href' => 'https://github.com/trk/WireMailPHPMailer',
             'author' => 'İskender TOTOĞLU | @ukyo(community), @trk (Github), https://www.altivebir.com',
@@ -71,18 +93,19 @@ class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule {
      *
      * WireMailPHPMailer constructor.
      */
-    public function __construct() {}
+    public function __construct()
+    {
+
+    }
 
     // ------------------------------------------------------------------------
 
     /**
      * Initialize the module
      */
-    public function init() {
-        $this->PHPMailer = new PHPMailer($this->exceptions);
-        foreach($this->getArray() as $key => $value) {
-            $this->PHPMailer->{$key} = $value;
-        }
+    public function init()
+    {
+
     }
 
     // ------------------------------------------------------------------------
@@ -92,7 +115,74 @@ class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule {
      *
      * @return PHPMailer
      */
-    public function mailer() { return $this->PHPMailer; }
+    public function mailer()
+    {
+        return $this->getInstance();
+    }
+
+    /**
+     * Return PHPMailer instance
+     *
+     * @return PHPMailer
+     */
+    public function getInstance($initialize = true)
+    {
+        $instance = new PHPMailer($this->exceptions);
+
+        if ($initialize) {
+            $instance = $this->initialize($instance);
+        }
+
+        return $instance;
+    }
+
+    /**
+     * Initialize module config
+     *
+     * @param PHPMailer|null $instance
+     * 
+     * @return PHPMailer|null
+     */
+    protected function initialize($instance = null)
+    {
+        if ($instance instanceof PHPMailer) {
+            foreach ($this->getArray() as $key => $value) {
+                $instance->{$key} = $value;
+            }
+        }
+
+        foreach ($this->options as $name => $value) {
+            if (in_array($name, ['AltBody', 'Subject', 'Body']) && is_string($value) && $value) {
+                $instance->{$name} = $value;
+            } else if (in_array($name, ['addAddress', 'addCC', 'addBCC', 'addReplyTo']) && is_array($value) && count($value)) {
+                foreach ($value as $e => $n) {
+                    $instance->{$name}($e, $n);
+                }
+            } else if ($name == 'setFrom' && $value['address']) {
+                $instance->setFrom($value['address'], $value['name'], $value['auto']);
+            } else if ($name == 'msgHTML' && $value['message']) {
+                $instance->msgHTML($value['message'], $value['basedir'], $value['advanced']);
+            } else if ($name == 'addAttachment' && is_array($value) && count($value)) {
+                foreach ($value as $p => $v) {
+                    $instance->addAttachment($p, $v['name'], $v['encoding'], $v['type'], $v['disposition']);
+                }
+            }
+        }
+
+        return $instance;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function Body($Body)
+    {
+        if (is_string($Body)) {
+            $this->options['Body'] = $Body;
+        }
+
+        return $this;
+    }
 
     // ------------------------------------------------------------------------
 
@@ -104,8 +194,12 @@ class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule {
      * @param string $AltBody
      * @return $this
      */
-    public function AltBody($AltBody="") {
-        if($AltBody) $this->PHPMailer->AltBody = $AltBody;
+    public function AltBody($AltBody="")
+    {
+        if ($AltBody) {
+            $this->options['AltBody'] = $AltBody;
+        }
+
         return $this;
     }
 
@@ -117,8 +211,12 @@ class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule {
      * @var string
      * @return $this
      */
-    public function addSubject($Subject="") {
-        if($Subject) $this->PHPMailer->Subject = $Subject;
+    public function addSubject($Subject="")
+    {
+        if ($Subject) {
+            $this->options['Subject'] = $Subject;
+        }
+        
         return $this;
     }
 
@@ -131,8 +229,10 @@ class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule {
      * @param string $name
      * @return $this
      */
-    public function addAddress($address, $name = '') {
-        $this->PHPMailer->addAddress($address, $name);
+    public function addAddress($address, $name = '')
+    {
+        $this->options['addAddress'][$address] = $name;
+        
         return $this;
     }
 
@@ -145,8 +245,10 @@ class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule {
      * @param string $name
      * @return $this
      */
-    public function addCC($address, $name = '') {
-        $this->PHPMailer->addCC($address, $name);
+    public function addCC($address, $name = '')
+    {
+        $this->options['addCC'][$address] = $name;
+        
         return $this;
     }
 
@@ -159,8 +261,10 @@ class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule {
      * @param string $name
      * @return $this
      */
-    public function addBCC($address, $name = '') {
-        $this->PHPMailer->addBCC($address, $name);
+    public function addBCC($address, $name = '')
+    {
+        $this->options['addBCC'][$address] = $name;
+        
         return $this;
     }
 
@@ -173,8 +277,10 @@ class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule {
      * @param string $name
      * @return $this
      */
-    public function addReplyTo($address, $name = '') {
-        $this->PHPMailer->addReplyTo($address, $name);
+    public function addReplyTo($address, $name = '')
+    {
+        $this->options['addReplyTo'][$address] = $name;
+        
         return $this;
     }
 
@@ -191,8 +297,9 @@ class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule {
      * @return array
      * @link http://www.andrew.cmu.edu/user/agreen1/testing/mrbs/web/Mail/RFC822.php A more careful implementation
      */
-    public function parseAddresses($addrstr, $useimap = true) {
-        return $this->PHPMailer::parseAddresses($addrstr, $useimap);
+    public function parseAddresses($addrstr, $useimap = true)
+    {
+        return PHPMailer::parseAddresses($addrstr, $useimap);
     }
 
     // ------------------------------------------------------------------------
@@ -206,8 +313,14 @@ class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule {
      * @return $this
      * @throws Exception
      */
-    public function setFrom($address, $name = '', $auto = true) {
-        $this->PHPMailer->setFrom($address, $name, $auto);
+    public function setFrom($address, $name = '', $auto = true)
+    {
+        $this->options['setFrom'] = [
+            'address' => $address,
+            'name' => $name,
+            'auto' => $auto
+        ];
+        
         return $this;
     }
 
@@ -221,9 +334,10 @@ class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule {
      *
      * @return string
      */
-    public function getLastMessageID() {
-        return $this->PHPMailer->getLastMessageID();
-    }
+    // public function getLastMessageID()
+    // {
+    //     return $this->PHPMailer->getLastMessageID();
+    // }
 
     // ------------------------------------------------------------------------
 
@@ -240,8 +354,14 @@ class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule {
      *    or your own custom converter @see PHPMailer::html2text()
      * @return string $message
      */
-    public function msgHTML($message, $basedir = '', $advanced = false) {
-        $this->PHPMailer->msgHTML($message, $basedir, $advanced);
+    public function msgHTML($message, $basedir = '', $advanced = false)
+    {
+        $this->options['msgHTML'] = [
+            'message' => $message,
+            'basedir' => $basedir,
+            'advanced' => $advanced
+        ];
+
         return $this;
     }
 
@@ -259,10 +379,19 @@ class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule {
      * @return $this
      * @throws Exception
      */
-    public function addAttachment($path, $name = '', $encoding = 'base64', $type = '', $disposition = 'attachment') {
-        $this->PHPMailer->addAttachment($path, $name, $encoding, $type, $disposition);
+    public function addAttachment($path, $name = '', $encoding = 'base64', $type = '', $disposition = 'attachment')
+    {
+        $this->options['addAttachment'][$path] = [
+            'name' => $name,
+            'encoding' => $encoding,
+            'type' => $type,
+            'disposition' => $disposition
+        ];
+        
         return $this;
     }
+
+    // ------------------------------------------------------------------------
 
     /**
      * Add a file to be attached to the email
@@ -275,10 +404,11 @@ class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule {
      * @return $this 
      *
      */
-    public function attachment($value, $filename = '') {
+    public function attachment($value, $filename = '')
+    {
         if($value === null) {
             // clear attachments
-            $this->PHPMailer->clearAttachments();
+            // $this->PHPMailer->clearAttachments();
         } else {
             // add attachment
             $this->addAttachment($value, $filename); 
@@ -295,31 +425,48 @@ class WireMailPHPMailer extends WireMail implements Module, ConfigurableModule {
      * @return int|void
      * @throws Exception
      */
-    public function ___send() {
+    public function ___send()
+    {
+
+        $instance = $this->getInstance();
+        
         if($this->mail["from"]) {
-            $this->PHPMailer->From = $this->mail["from"];
-            $this->PHPMailer->FromName = $this->mail["fromName"];
+            $instance->From = $this->mail["from"];
+            $instance->FromName = $this->mail["fromName"];
         }
-        if($this->mail["subject"]) $this->PHPMailer->Subject = $this->mail["subject"];
+
+        if($this->mail["subject"]) {
+            $instance->Subject = $this->mail["subject"];
+        }
+        
         foreach($this->to as $to) {
             $toName = $this->mail['toName'][$to];
-            $this->PHPMailer->addAddress($to, $toName);
+            $instance->addAddress($to, $toName);
         }
+
         if(count($this->attachments)) {
             foreach($this->attachments as $filename => $file) {
-                $this->PHPMailer->addAttachment($file, $filename);
+                $instance->addAttachment($file, $filename);
             }
         }
+
         if($this->mail["replyTo"]) {
-            $this->PHPMailer->addReplyTo($this->mail["replyTo"], $this->mail["replyToName"]);
+            $instance->addReplyTo($this->mail["replyTo"], $this->mail["replyToName"]);
         }
+
         if($this->bodyHTML) {
-            $this->PHPMailer->isHTML(true);
-            $this->PHPMailer->Body = $this->bodyHTML;
+            $instance->isHTML(true);
+            $instance->Body = $this->bodyHTML;
         }
+
         if($this->body) {
-            $this->PHPMailer->AltBody = $this->body;
+            $instance->AltBody = $this->body;
         }
-        return $this->PHPMailer->send() ? 1 : 0;
+        bd($instance);
+        if (! $instance->send()) {
+            return false;
+        }
+
+        return true;
     }
 }